Static buffered geometry and drawers for spheres and cones.
+++ gl-buffered.js contains a class for geometry corresponding to a
GL buffer. This is a static buffer (using STATIC_DRAW) and does
not allow data modification after the buffer is finalized (which
happens upon first drawing it).

TODO: add similar dynamic and/or stream buffers with similar API's.

+++ drawers.js contains objects for creating and drawing geometry.
For instance, calling SphereDrawer creates the vertex data for a
sphere, puts it into a GL buffer and returns a corresponding
GLBuffered object.