Matrix Parametrized Transformations (MPT's).
From Library Comments:
The objects in this library are useful for animation (transformations
parametrized with time, e.g.). The library provides a series of factories
for different transformation types (e.g. MPTRotateX is a rotation
about the X-axis with a single degree parameter). It is not advised
to call directly onto MPT types (e.g. CachedMPT, ComboMPT, etc.), but
to use factories instead.